Ported APM to ts
