--- src/lib/apm/index.js
@@ -1,5 +1,0 @@
-const server = require('./ServerSide');
-const client = require('./ClientSide');
-const apm = typeof window === 'undefined' ? server : client;
-
-module.exports = apm;
